# docs/swagger/swagger.yaml
basePath: /v1
definitions:
...
  dto.CustomerInvoiceSummary:
    properties:
      currency:
        description: currency is the three-letter ISO currency code for this summary
        type: string
      customer_id:
        description: customer_id is the unique identifier of the customer
        type: string
      overdue_invoice_count:
        description: overdue_invoice_count is the number of overdue invoices for this
          customer in this currency
        type: integer
      total_invoice_count:
        description: total_invoice_count is the total number of invoices for this
          customer in this currency
        type: integer
      total_overdue_amount:
        description: total_overdue_amount is the total amount of overdue invoices
          in this currency
        type: number
      total_revenue_amount:
        description: total_revenue_amount is the total revenue generated from this
          customer in this currency
        type: number
      total_unpaid_amount:
        description: total_unpaid_amount is the total amount of unpaid invoices in
          this currency
        type: number
      unpaid_fixed_charges:
        description: unpaid_fixed_charges is the total amount of unpaid fixed charges
          in this currency
        type: number
      unpaid_invoice_count:
        description: unpaid_invoice_count is the number of unpaid invoices for this
          customer in this currency
        type: integer
      unpaid_usage_charges:
        description: unpaid_usage_charges is the total amount of unpaid usage-based
          charges in this currency
        type: number
    type: object
  dto.CustomerMultiCurrencyInvoiceSummary:
    properties:
      customer_id:
        description: customer_id is the unique identifier of the customer
        type: string
      default_currency:
        description: default_currency is the primary currency for this customer
        type: string
      summaries:
        description: summaries contains the invoice summaries for each currency
        items:
          $ref: '#/definitions/dto.CustomerInvoiceSummary'
        type: array
    type: object
...
  dto.GetPreviewInvoiceRequest:
    properties:
      period_end:
        description: period_end is the optional end date of the period to preview
        type: string
      period_start:
        description: period_start is the optional start date of the period to preview
        type: string
      subscription_id:
        description: subscription_id is the unique identifier of the subscription
          to preview invoice for
        type: string
    required:
    - subscription_id
    type: object
...
  dto.ListInvoicesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.InvoiceResponse'
        type: array
      pagination:
        $ref: '#/definitions/types.PaginationResponse'
    type: object
...
  dto.UpdateInvoiceRequest:
    properties:
      invoice_pdf_url:
        description: invoice_pdf_url is the URL where customers can download the PDF
          version of this invoice
        type: string
    type: object
...
  dto.UpdatePaymentStatusRequest:
    properties:
      amount:
        description: amount is the optional payment amount to record
        type: number
      payment_status:
        allOf:
        - $ref: '#/definitions/types.PaymentStatus'
        description: payment_status is the new payment status to set for the invoice
          (paid, unpaid, etc.)
    required:
    - payment_status
    type: object
...
  errors.ErrorDetail:
    properties:
      details:
        additionalProperties: {}
        type: object
      internal_error:
        type: string
      message:
        type: string
    type: object
  errors.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/errors.ErrorDetail'
      success:
        type: boolean
    type: object
  gin.H:
    additionalProperties: {}
    type: object
...
  types.PaginationResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
info:
  contact: {}
  description: FlexPrice API Service
  title: FlexPrice API
  version: "1.0"
paths:
...
  /customers/{id}/invoices/summary:
    get:
      consumes:
      - application/json
      description: Get a customer invoice summary
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CustomerMultiCurrencyInvoiceSummary'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a customer invoice summary
      tags:
      - Invoices
...
  /invoices:
    get:
      consumes:
      - application/json
      description: List invoices with optional filtering
      parameters:
      - description: |-
          amount_due_gt filters invoices with a total amount due greater than the specified value
          Useful for finding invoices above a certain threshold or identifying high-value invoices
        in: query
        name: amount_due_gt
        type: number
      - description: |-
          amount_remaining_gt filters invoices with an outstanding balance greater than the specified value
          Useful for finding invoices that still have significant unpaid amounts
        in: query
        name: amount_remaining_gt
        type: number
      - description: |-
          customer_id filters invoices for a specific customer using FlexPrice's internal customer ID
          This is the ID returned by FlexPrice when creating or retrieving customers
        in: query
        name: customer_id
        type: string
      - in: query
        name: end_time
        type: string
      - in: query
        name: expand
        type: string
      - description: |-
          external_customer_id filters invoices for a customer using your system's customer identifier
          This is the ID you provided when creating the customer in FlexPrice
        in: query
        name: external_customer_id
        type: string
      - collectionFormat: csv
        description: |-
          invoice_ids restricts results to invoices with the specified IDs
          Use this to retrieve specific invoices when you know their exact identifiers
        in: query
        items:
          type: string
        name: invoice_ids
        type: array
      - collectionFormat: csv
        description: |-
          invoice_status filters by the current state of invoices in their lifecycle
          Multiple statuses can be specified to include invoices in any of the listed states
        in: query
        items:
          enum:
          - DRAFT
          - FINALIZED
          - VOIDED
          type: string
        name: invoice_status
        type: array
      - description: |-
          invoice_type filters by the nature of the invoice (SUBSCRIPTION, ONE_OFF, or CREDIT)
          Use this to separate recurring charges from one-time fees or credit adjustments
        enum:
        - SUBSCRIPTION
        - ONE_OFF
        - CREDIT
        in: query
        name: invoice_type
        type: string
        x-enum-varnames:
        - InvoiceTypeSubscription
        - InvoiceTypeOneOff
        - InvoiceTypeCredit
      - in: query
        maximum: 1000
        minimum: 1
        name: limit
        type: integer
      - in: query
        minimum: 0
        name: offset
        type: integer
      - enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - collectionFormat: csv
        description: |-
          payment_status filters by the payment state of invoices
          Multiple statuses can be specified to include invoices with any of the listed payment states
        in: query
        items:
          enum:
          - INITIATED
          - PENDING
          - PROCESSING
          - SUCCEEDED
          - FAILED
          - REFUNDED
          - PARTIALLY_REFUNDED
          type: string
        name: payment_status
        type: array
      - in: query
        name: start_time
        type: string
      - enum:
        - published
        - deleted
        - archived
        in: query
        name: status
        type: string
        x-enum-varnames:
        - StatusPublished
        - StatusDeleted
        - StatusArchived
      - description: |-
          subscription_id filters invoices generated for a specific subscription
          Only returns invoices that were created as part of the specified subscription's billing
        in: query
        name: subscription_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListInvoicesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: List invoices
      tags:
      - Invoices
    post:
      consumes:
      - application/json
      description: Create a new invoice with the provided details
      parameters:
      - description: Invoice details
        in: body
        name: invoice
        required: true
        schema:
          $ref: '#/definitions/dto.CreateInvoiceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.InvoiceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Create a new invoice
      tags:
      - Invoices
  /invoices/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about an invoice
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InvoiceResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get an invoice by ID
      tags:
      - Invoices
    put:
      consumes:
      - application/json
      description: Update invoice details like PDF URL
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      - description: Invoice Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateInvoiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InvoiceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Update an invoice
      tags:
      - Invoices
  /invoices/{id}/finalize:
    post:
      consumes:
      - application/json
      description: Finalize a draft invoice
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Finalize an invoice
      tags:
      - Invoices
  /invoices/{id}/payment:
    put:
      consumes:
      - application/json
      description: Update the payment status of an invoice
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      - description: Payment Status Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePaymentStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InvoiceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update invoice payment status
      tags:
      - Invoices
  /invoices/{id}/payment/attempt:
    post:
      consumes:
      - application/json
      description: Attempt to pay an invoice using customer's available wallets
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Attempt payment for an invoice
      tags:
      - Invoices
  /invoices/{id}/pdf:
    get:
      description: Retrieve the PDF document for a specific invoice by its ID
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      - description: Return presigned URL from s3 instead of PDF
        in: query
        name: url
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get PDF for an invoice
      tags:
      - Invoices
  /invoices/{id}/recalculate:
    post:
      consumes:
      - application/json
      description: Recalculate totals and line items for a draft invoice, useful when
        subscription line items or usage data has changed
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Whether to finalize the invoice after recalculation (default:
          true)'
        in: query
        name: finalize
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InvoiceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Recalculate invoice totals and line items
      tags:
      - Invoices
  /invoices/{id}/void:
    post:
      consumes:
      - application/json
      description: Void an invoice that hasn't been paid
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Void an invoice
      tags:
      - Invoices
  /invoices/preview:
    post:
      consumes:
      - application/json
      description: Get a preview invoice
      parameters:
      - description: Preview Invoice Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GetPreviewInvoiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InvoiceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get a preview invoice
      tags:
      - Invoices
  /invoices/search:
    post:
      consumes:
      - application/json
      description: List invoices by filter
      parameters:
      - description: Filter
        in: body
        name: filter
        required: true
        schema:
          $ref: '#/definitions/types.InvoiceFilter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListInvoicesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List invoices by filter
      tags:
      - Invoices
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Enter your API key in the format *x-api-key &lt;api-key&gt;**
    in: header
    name: x-api-key
    type: apiKey
swagger: "2.0"